add support for Lights (ZwaveDimmer) (#36)
* add support for Lights (ZwaveDimmer)

Basic light support (non-color dimmer only). Color support to be added later.

Also a includes a small fix for displaying the power_consumption as  state attribute if a device supports this.

* remove power_consumption state attribute

power_consumption was historically added as state attribute to devices that supported it. Now it has it's own sensor (device_electricity_w)

* follow up fixes

prefer target over level if exists on light
fix a few typos and styling stuff

* add "in" operator

to lookup value exists in entity values

* lookup value in dict with get